// app/enums/spacex.enum.ts
export enum ProcessStatus {
    IDLE = 'idle',
    LOADING = 'loading',
    SUCCESS = 'success',
    ERROR = 'error',
}

export enum Routes {
    Dashboard = "/(dashboard)",
    Launches = "/launches",
    Favorites = "/favorites",
}


export enum LaunchStatus {
    All = 'All',
    Success = 'Success',
    Failure = 'Failure',
    Upcoming = 'Upcoming',
}